CheckUpdateButton is a custom view and Butterknife
The checkupdate button is now implemented as a custom view, this means
that is configurable in the layout's xml as an Android default View.

This should increase code readability and maintainability.

ButterKnife is now used to get Views from layouts and to perfrom other
Android framework view tasks.

This should increase the code quality.