Correct misalignement of class and .java name
The name of .java for StretchedBackgroundView class wasn't the same as the
class.

The class worked as intended in a debug apk because its
visibility was set to default (not public nor private) and the compiler
interpred it as a nested class inside the .java. However, when
generating the signed apk, the deafult visibility of the class caused an
error.

Renaming the name of the .java with the same name of the class and
setting the visibility of the latter to public fixes the problem.